+++ partiql-types/src/main/kotlin/org/partiql/types/StaticType.kt
@@ -464,8 +464,17 @@ public data class BagType(
     override fun toString(): String = "bag($elementType)"
 }
 
+/**
+ * Describes a PartiQL Struct.
+ *
+ * @param fields the key-value pairs of the struct
+ * @param contentClosed when true, denotes that no other attributes may be present
+ * @param primaryKeyFields fields designated as primary keys
+ * @param constraints set of constraints applied to the Struct
+ * @param metas meta-data
+ */
 public data class StructType(
-    val fields: Map<String, StaticType> = mapOf(),
+    val fields: List<Field> = listOf(),
     // `TupleConstraint` already has `Open` constraint which overlaps with `contentClosed`.
     // In addition, `primaryKeyFields` must not exist on the `StructType` as `PrimaryKey`
     // is a property of collection of tuples. As we have plans to define PartiQL types in
@@ -479,17 +488,45 @@ public data class StructType(
     val constraints: Set<TupleConstraint> = setOf(),
     override val metas: Map<String, Any> = mapOf(),
 ) : SingleType() {
+
+    public constructor(
+        fields: Map<String, StaticType>,
+        contentClosed: Boolean = false,
+        primaryKeyFields: List<String> = listOf(),
+        constraints: Set<TupleConstraint> = setOf(),
+        metas: Map<String, Any> = mapOf(),
+    ) : this(
+        fields.map { Field(it.key, it.value) },
+        contentClosed,
+        primaryKeyFields,
+        constraints,
+        metas
+    )
+
+    /**
+     * The key-value pair of a StructType, where the key represents the name of the field and the value represents
+     * its [StaticType]. Note: multiple [Field]s within a [StructType] may contain the same [key], and therefore,
+     * multiple same-named keys may refer to distinct [StaticType]s. To determine the [StaticType]
+     * of a reference to a field, especially in the case of duplicates, it depends on the ordering of the [StructType]
+     * (denoted by the presence of [TupleConstraint.Ordered] in the [StructType.constraints]).
+     * - If ORDERED: the PartiQL specification says to grab the first encountered matching field.
+     * - If UNORDERED: it is implementation-defined. However, gather all possible types, merge them using [AnyOfType].
+     */
+    data class Field(
+        val key: String,
+        val value: StaticType
+    )
+
     override fun flatten(): StaticType = this
 
     override val allTypes: List<StaticType>
         get() = listOf(this)
 
     override fun toString(): String {
-        val entries = fields.entries
-        val firstSeveral = entries.toList().take(3).joinToString { "${it.key}: ${it.value}" }
+        val firstSeveral = fields.take(3).joinToString { "${it.key}: ${it.value}" }
         return when {
-            entries.size <= 3 -> "struct($firstSeveral, $constraints)"
-            else -> "struct($firstSeveral, ... and ${entries.size - 3} other field(s), $constraints)"
+            fields.size <= 3 -> "struct($firstSeveral, $constraints)"
+            else -> "struct($firstSeveral, ... and ${fields.size - 3} other field(s), $constraints)"
         }
     }
 }
@@ -574,6 +611,12 @@ public sealed class NumberConstraint {
 public sealed class TupleConstraint {
     public data class UniqueAttrs(val value: Boolean) : TupleConstraint()
     public data class Open(val value: Boolean) : TupleConstraint()
+
+    /**
+     * The presence of the [Ordered] on a [StructType] represents that the [StructType] is ORDERED. The absence of
+     * this constrain represents the opposite -- AKA that the [StructType] is UNORDERED
+     */
+    public object Ordered : TupleConstraint()
 }
 
 /**